DeepLearning/DeployCifar10/requirements.txt
absl-py==1.0.0
astunparse==1.6.3
cached-property==1.5.2
cachetools==4.2.4
certifi==2021.5.30
charset-normalizer==2.0.10
click==8.0.1
colorama==0.4.4
Flask==2.0.1
flatbuffers==2.0
gast==0.4.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.43.0
gunicorn==20.1.0
h5py==3.6.0
idna==3.3
importlib-metadata==4.6.1
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
keras==2.7.0
Keras-Preprocessing==1.1.2
libclang==12.0.0
Markdown==3.3.6
MarkupSafe==2.0.1
numpy==1.21.1
oauthlib==3.1.1
opencv-python==4.5.5.62
opt-einsum==3.3.0
pandas==1.1.5
Pillow==9.0.0
protobuf==3.19.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.2
pytz==2021.1
requests==2.27.1
requests-oauthlib==1.3.0
rsa==4.8
scikit-learn==0.24.2
scipy==1.7.0
six==1.16.0
sklearn==0.0
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.23.1
termcolor==1.1.0
threadpoolctl==2.2.0
typing-extensions==3.10.0.0
urllib3==1.26.8
Werkzeug==2.0.1
wincertstore==0.2
wrapt==1.13.3
zipp==3.5.0
